Two Matrix Action
A flaw was found in the matrix action when the determinant of H was not -1, 0, or 1. The mathematics behind the attack is described in Section 6.2 of the paper.

To remedy this, the protocol now calls for the generation of two matrices, H_1 and H_2, that both have determinant 0. A quick walkthrough of how the generation of these matrices occurs:
       - A diagonal matrix D is generated, where the first diagonal entry is 0, and the rest are numbers mod p (note that because the matrix is set to a 3 x 3 matrix, only two values are generated). Therefore, the determinant is 0.
       - A matrix S is generated with random values. When it fits a check to see if it is invertible (determinant is not 0), we find the inverse of this matrix and call it S_INV.
       - Then, H = S_INV * D * S

The bulk of this update involved creating a function to invert a matrix. I created a function to find the LU decomposition of a matrix mod p (LUDecomp), and another to solve a system of linear equations under mod p (LUSolve). The key difference between LU decompositions under mod p and for normal matrices is the division of the pivot being replaced with a multiplication under mod p.

The rest of the procedure remains mostly the same. Note that with two functions to generate matrices, the actual protocol is much more readable.